+++ interpreter/k8s/src/test/scala/io/buoyant/transformer/k8s/LocalnodeTransformerTest.scala
@@ -12,7 +12,7 @@ import java.net.{InetAddress, InetSocketAddress}
 class LocalnodeTransformerTest extends FunSuite {
 
   test("localnode transformer selects local instance") {
-    val transformer = new SubnetLocalTransformer(InetAddress.getByName("1.1.1.1"), Netmask("255.255.255.0"))
+    val transformer = new SubnetLocalTransformer(Seq(InetAddress.getByName("1.1.1.1")), Netmask("255.255.255.0"))
 
     val interpreter = new NameInterpreter {
       override def bind(