+++ interpreter/per-host/src/main/scala/io/buoyant/transformer/perHost/LocalhostTransformerInitializer.scala
@@ -2,7 +2,8 @@ package io.buoyant.transformer
 package perHost
 
 import io.buoyant.namer.{NameTreeTransformer, TransformerConfig, TransformerInitializer}
-import java.net.InetAddress
+import java.net.NetworkInterface
+import scala.collection.JavaConverters._
 
 class LocalhostTransformerInitializer extends TransformerInitializer {
   val configClass = classOf[LocalhostTransformerConfig]
@@ -11,6 +12,14 @@ class LocalhostTransformerInitializer extends TransformerInitializer {
 
 class LocalhostTransformerConfig extends TransformerConfig {
 
-  override def mk(): NameTreeTransformer =
-    new SubnetLocalTransformer(InetAddress.getLocalHost, Netmask("255.255.255.255"))
+  override def mk(): NameTreeTransformer = {
+    val localIPs = for {
+      interface <- NetworkInterface.getNetworkInterfaces.asScala
+      if interface.isUp
+      inet <- interface.getInetAddresses.asScala
+      if !inet.isLoopbackAddress
+    } yield inet
+    new SubnetLocalTransformer(localIPs.toSeq, Netmask("255.255.255.255"))
+  }
+
 }